# openapi/fountain-life-api.json
{
  "openapi": "3.0.0",
  "paths": {
    "/accounts/get-current-account": {
      "post": {
        "operationId": "getCurrentAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/register-current-account": {
      "post": {
        "operationId": "registerCurrentAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAccountResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "accounts"
        ]
      }
    },
    "/assistants/list-assistants": {
      "post": {
        "operationId": "listAssistants",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListAssistantsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAssistantsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assistants"
        ]
      }
    },
    "/assistants/get-conversation": {
      "post": {
        "operationId": "getConversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetAssistantConversationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAssistantConversationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assistants"
        ]
      }
    },
    "/assistants/{assistantKey}/stream-message": {
      "post": {
        "operationId": "streamAssistantMessage",
        "parameters": [
          {
            "name": "assistantKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendAssistantMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server-sent event stream. Each data event is an AssistantThreadUpdateResponse.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantThreadUpdateResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "assistants"
        ]
      }
    },
    "/documents/upload-document": {
      "post": {
        "operationId": "uploadDocument",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadDocumentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/list-documents": {
      "post": {
        "operationId": "listDocuments",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListDocumentsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDocumentsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/delete-document": {
      "post": {
        "operationId": "deleteDocument",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDocumentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
...
    "/mcp": {
      "post": {
        "operationId": "handleMcpRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC 2.0 response for the MCP streamable HTTP transport."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "mcp"
        ]
      }
    }
  },
  "info": {
    "title": "Fountain Life Notebook API",
    "description": "NestJS API for the Fountain Life interview NotebookLM-style app.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "BaseRequest": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "payload": {
            "type": "object"
          }
        }
      },
      "ErrorInfo": {
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        },
        "required": [
          "errorMessage"
        ]
      },
      "AuthenticatedUserInfo": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "subject": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "subject",
          "username"
        ]
      },
      "CurrentUserResponse": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            }
          },
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AuthenticatedUserInfo"
          }
        },
        "required": [
          "errors",
          "isSuccess"
        ]
      },
      "AccountSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "cognitoSubject": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "cognitoSubject",
          "email",
          "username"
        ]
      },
      "RegisterAccountResponse": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            }
          },
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AccountSummary"
          }
        },
        "required": [
          "errors",
          "isSuccess"
        ]
      },
      "ListAssistantsRequest": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "payload": {
            "type": "object"
          }
        }
      },
      "AssistantSummary": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "key",
          "name"
        ]
      },
      "ListAssistantsResponse": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            }
          },
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssistantSummary"
            }
          }
        },
        "required": [
          "errors",
          "isSuccess"
        ]
      },
      "GetAssistantConversationPayload": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string"
          }
        },
        "required": [
          "conversationId"
        ]
      },
      "GetAssistantConversationRequest": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/GetAssistantConversationPayload"
          }
        }
      },
      "Citation": {
        "type": "object",
        "properties": {
          "chunkIndex": {
            "type": "number"
          },
          "documentId": {
            "type": "string"
          },
          "documentName": {
            "type": "string"
          },
          "snippet": {
            "type": "string"
          }
        },
        "required": [
          "chunkIndex",
          "documentId",
          "documentName",
          "snippet"
        ]
      },
      "AssistantConversationMessage": {
        "type": "object",
        "properties": {
          "actorDisplayName": {
            "type": "string"
          },
          "actorUserId": {
            "type": "string"
          },
          "citations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Citation"
            }
          },
          "createdDateUtc": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "assistant",
              "system",
              "tool",
              "user"
            ]
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "createdDateUtc",
          "id",
          "role",
          "text"
        ]
      },
      "AssistantConversationParticipant": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "joinedDateUtc": {
            "format": "date-time",
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "member",
              "owner"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "removed"
            ]
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "joinedDateUtc",
          "role",
          "status",
          "userId"
        ]
      },
      "AssistantConversation": {
        "type": "object",
        "properties": {
          "assistantKey": {
            "type": "string"
          },
          "createdDateUtc": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssistantConversationMessage"
            }
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssistantConversationParticipant"
            }
          },
          "lastUpdatedDateUtc": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "assistantKey",
          "createdDateUtc",
          "id",
          "messages",
          "participants",
          "lastUpdatedDateUtc"
        ]
      },
      "GetAssistantConversationResponse": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            }
          },
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AssistantConversation"
          }
        },
        "required": [
          "errors",
          "isSuccess"
        ]
      },
      "SendAssistantMessagePayload": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          },
          "documentIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "participantUserIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "message"
        ]
      },
      "SendAssistantMessageRequest": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/SendAssistantMessagePayload"
          }
        }
      },
      "AssistantThreadUpdate": {
        "type": "object",
        "properties": {
          "citations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Citation"
            }
          },
          "conversationId": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "assistant",
              "system",
              "user"
            ]
          },
          "text": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "message",
              "status"
            ]
          }
        },
        "required": [
          "conversationId",
          "role",
          "text",
          "type"
        ]
      },
      "AssistantThreadUpdateResponse": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            }
          },
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AssistantThreadUpdate"
          }
        },
        "required": [
          "errors",
          "isSuccess"
        ]
      },
      "DocumentSummary": {
        "type": "object",
        "properties": {
          "byteSize": {
            "type": "number"
          },
          "chunkCount": {
            "type": "number"
          },
          "contentType": {
            "type": "string"
          },
          "createdDateUtc": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lastUpdatedDateUtc": {
            "format": "date-time",
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "failed",
              "ready"
            ]
          },
          "textPreview": {
            "type": "string"
          }
        },
        "required": [
          "byteSize",
          "chunkCount",
          "contentType",
          "createdDateUtc",
          "id",
          "lastUpdatedDateUtc",
          "originalFileName",
          "status"
        ]
      },
      "UploadDocumentResponse": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            }
          },
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DocumentSummary"
          }
        },
        "required": [
          "errors",
          "isSuccess"
        ]
      },
      "ListDocumentsRequest": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "payload": {
            "type": "object"
          }
        }
      },
      "ListDocumentsResponse": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            }
          },
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentSummary"
            }
          }
        },
        "required": [
          "errors",
          "isSuccess"
        ]
      },
      "DeleteDocumentPayload": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string"
          }
        },
        "required": [
          "documentId"
        ]
      },
      "DeleteDocumentRequest": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/DeleteDocumentPayload"
          }
        }
      },
      "DeleteDocumentResult": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ]
      },
      "DeleteDocumentResponse": {
        "type": "object",
        "properties": {
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorInfo"
            }
          },
          "isSuccess": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DeleteDocumentResult"
          }
        },
        "required": [
          "errors",
          "isSuccess"
        ]
      }
    }
  }
}